ddd-apps/requirements.txt
flask==3.0.2
openai==1.12.0
selenium==4.18.1
python-dotenv==1.0.1 